Refactored Project and Navigation Structure
