+++ app/src/main/java/com/rf/bandmeets/navigation/Navigation.kt
@@ -1,0 +1,26 @@
+package com.rf.bandmeets.navigation
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import com.rf.bandmeets.ui.screens.HomeScreen
+import com.rf.bandmeets.ui.screens.Screen
+import com.rf.bandmeets.ui.screens.SplashScreen
+
+@Composable
+fun Navigation() {
+    val navController = rememberNavController()
+    NavHost(
+        navController = navController,
+        startDestination = Screen.SplashScreen.route
+    ) {
+        composable(route = Screen.SplashScreen.route) {
+            SplashScreen(navController = navController)
+        }
+        composable(route = Screen.MainScreen.route) {
+            HomeScreen(navController = navController)
+        }
+    }
+
+}
