+++ app/src/main/java/com/rf/bandmeets/ui/screens/Screen.kt
@@ -1,4 +1,4 @@
-package com.rf.bandmeets
+package com.rf.bandmeets.ui.screens
 
 sealed class Screen(val route: String){
     object MainScreen: Screen("main_screen")
